Create user profile route, update all links to user profile to correct route
