+++ client/src/pages/DashboardPage.module.scss
@@ -1,25 +1,3 @@
-.header {
-  background-color: #202020;
-  display: flex;
-  align-items: center;
-  height: 4rem;
-  padding: 1rem;
-  justify-content: space-between;
-  color: #B0AEAA;
-
-  & > section {
-    display: flex;
-    align-items: center;
-    gap: 1.5rem;
-  }
-
-  h1 {
-    color: #FFA115;
-    font-size: 1rem;
-    font-weight: 500;
-  }
-}
-
 .avatar {
   height: 2rem;
   width: 2rem;
@@ -54,28 +32,37 @@
     text-align: center;
     grid-area: stats;
 
-    .userInfo {
-      display: grid;
-      gap: 0.25rem;
-      margin-top: 0.75rem;
-      grid-template-areas: 
-        "name badge"
-        "username username";
-      
-      h2 {
-        color: #fff;
-        font-size: 1rem;
-        grid-area: name;
-      }
+    a {
+      text-decoration: none;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      color: #B0AEAA;
 
-      svg {
-        grid-area: badge;
-      }
-
-      p {
-        grid-area: username;
+      .userInfo {
+        display: grid;
+        gap: 0.25rem;
+        margin-top: 0.75rem;
+        grid-template-areas: 
+          "name badge"
+          "username username";
+        
+        h2 {
+          color: #fff;
+          font-size: 1rem;
+          grid-area: name;
+        }
+  
+        svg {
+          grid-area: badge;
+        }
+  
+        p {
+          grid-area: username;
+        }
       }
     }
+
 
     .statsGrid {
       display: grid;
@@ -163,6 +150,7 @@
   padding: 1.5rem;
   display: grid;
   gap: 2rem;
+  color: #B0AEAA;
 
   button {
     display: flex;
@@ -186,11 +174,19 @@
     justify-content: unset;
     padding: unset;
     flex-wrap: wrap;
+    align-items: center;
 
-    h3 {
-      font-size: 1rem;
-      font-weight: 600;
-      color: #fff;
+    a {
+      display: flex;
+      align-items: center;
+      gap: 0.75rem;
+      text-decoration: none;
+
+      h3 {
+        font-size: 1rem;
+        font-weight: 600;
+        color: #fff;
+      }
     }
 
     .detailText {
@@ -238,7 +234,8 @@
       border-top: 1px solid #2C2C2C;
     }
 
-    li {
+    li a {
+      text-decoration: none;
       display: flex;
       gap: 0.75rem;
       padding: 1rem;
@@ -271,13 +268,4 @@
 
     }
   }
-}
-
-.btnText {
-  background: none;
-  border: none;
-  width: fit-content;
-  cursor: pointer;
-  color: #B0AEAA;
-  font-weight: 600;
 }
